Fix resource leak in server.
There is a problem with the tight loop that waits until the connection
is closed. connection.close? will never be true since we haven't closed
the connection. When the client tears down the TCP connection
gracefully, we will recieve a FIN. This doesn't close the socket,
however. It generates an EOF condition, which we must check for.